Security: PGN-Reader: OOB-Read bei signed-char-Indizierung beheben (b22)
isMoveToken() und skipMoveToken() indizieren CharToType[int(*s)]. Bei
signed char (Default auf x86) wird ein Byte >= 0x80 (non-ASCII, in PGN
mit Latin-1/UTF-8 haeufig) negativ -> CharToType[negativ] liest bis zu
128 Byte VOR dem 256-Byte-Array (Out-of-Bounds-Read, UB). Erreichbar
ueber ein PGN mit non-ASCII-Byte in einem ungueltigen Move-Token
(Fehlerpfad bei sendError, Zeile ~4610).

Fix: Byte(*s) statt int(*s) (Byte == unsigned char), konsistent mit den
uebrigen CharToType-Zugriffen (z.B. Zeile 5500ff).

Nur Lese-Zugriff in gemapptes rodata -> kein Crash/Write, geringe
Schwere, aber UB beseitigt.

+++ src/db/db_pgn_reader.cpp
@@ -239,7 +239,7 @@ quote(mstl::string const& s)
 static bool
 isMoveToken(char const* s)
 {
-	while (CharToType[int(*s)] != None)
+	while (CharToType[Byte(*s)] != None)
 		++s;
 
 	return !::isalnum(*s);
@@ -249,7 +249,7 @@ isMoveToken(char const* s)
 static char const*
 skipMoveToken(char const* s)
 {
-	while (CharToType[int(*s)] != None)
+	while (CharToType[Byte(*s)] != None)
 		++s;
 
 	return s;
